// src/lib/components/types.ts
type Radius = 'none' | 'sm' | 'md' | 'lg' | 'full';
type Size = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'auto';
type Color = 'black'
    | 'white'
    | 'gray'
    | 'none'
    | 'default'
    | 'primary'
    | 'secondary'
    | 'success'
    | 'warning'
    | 'danger';
type Variant = 'solid' | 'bordered' | 'shadow';
type Underline = 'none' | 'hover' | 'always';
type Lang = 'js' | 'ts' | 'css' | 'py' | 'bash';

export type { Radius, Size, Color, Variant, Underline, Lang }